test: migrate 'create presentation'
