+++ packages/core/src/credential/Credential.spec.ts
@@ -13,7 +13,6 @@ import { encodeAddress } from '@polkadot/util-crypto'
 
 import type {
   CompressedCredential,
-  DidKey,
   DidResolvedDetails,
   DidUri,
   IAttestation,
@@ -37,7 +36,6 @@ import {
   KeyTool,
   makeSigningKeyTool,
 } from '@kiltprotocol/testing'
-import { UUID } from '@kiltprotocol/utils'
 import * as Attestation from '../attestation'
 import * as Claim from '../claim'
 import * as CType from '../ctype'
@@ -87,27 +85,6 @@ async function buildCredential(
     attesterDid
   )
   return [credential, testAttestation]
-}
-
-// Returns a full DID that has the same identifier of the first light DID, but the same key authentication key as the second one, if provided, or as the first one otherwise.
-function createMinimalFullDidFromLightDid(
-  lightDidForId: LightDidDetails,
-  newAuthenticationKey?: DidKey
-): FullDidDetails {
-  const uri = DidUtils.getKiltDidFromIdentifier(
-    lightDidForId.identifier,
-    'full'
-  )
-  const authKey = newAuthenticationKey || lightDidForId.authenticationKey
-
-  return new FullDidDetails({
-    identifier: lightDidForId.identifier,
-    uri,
-    keyRelationships: {
-      authentication: new Set([authKey.id]),
-    },
-    keys: { [authKey.id]: authKey },
-  })
 }
 
 describe('Credential', () => {
@@ -348,312 +325,5 @@ describe('Credential', () => {
       keyAlice.sign
     )
     expect(Credential.getHash(credential)).toEqual(attestation.claimHash)
-  })
-})
-
-describe('create presentation', () => {
-  let migratedClaimerLightDid: DidDetails
-  let migratedClaimerFullDid: DidDetails
-  let newKeyForMigratedClaimerDid: KeyTool
-  let unmigratedClaimerLightDid: DidDetails
-  let unmigratedClaimerKey: KeyTool
-  let migratedThenDeletedClaimerLightDid: DidDetails
-  let migratedThenDeletedKey: KeyTool
-  let migratedThenDeletedClaimerFullDid: DidDetails
-  let attester: DidDetails
-  let ctype: ICType
-  let reqForAtt: ICredential
-
-  const mockResolver: IDidResolver = (() => {
-    const resolve = async (
-      didUri: DidUri
-    ): Promise<DidResolvedDetails | null> => {
-      // For the mock resolver, we need to match the base URI, so we delete the fragment, if present.
-      const { did } = DidUtils.parseDidUri(didUri)
-      switch (did) {
-        case migratedClaimerLightDid?.uri:
-          return {
-            details: migratedClaimerLightDid,
-            metadata: {
-              canonicalId: migratedClaimerFullDid.uri,
-              deactivated: false,
-            },
-          }
-        case migratedThenDeletedClaimerLightDid?.uri:
-          return {
-            metadata: {
-              deactivated: true,
-            },
-          }
-        case migratedThenDeletedClaimerFullDid?.uri:
-          return {
-            metadata: {
-              deactivated: true,
-            },
-          }
-        case unmigratedClaimerLightDid?.uri:
-          return {
-            details: unmigratedClaimerLightDid,
-            metadata: { deactivated: false },
-          }
-        case migratedClaimerFullDid?.uri:
-          return {
-            details: migratedClaimerFullDid,
-            metadata: { deactivated: false },
-          }
-        case attester?.uri:
-          return { details: attester, metadata: { deactivated: false } }
-        default:
-          return null
-      }
-    }
-    return {
-      resolve,
-      resolveDoc: resolve,
-    } as IDidResolver
-  })()
-
-  beforeAll(async () => {
-    const { keypair } = makeSigningKeyTool()
-    attester = await createLocalDemoFullDidFromKeypair(keypair)
-
-    unmigratedClaimerKey = makeSigningKeyTool()
-    unmigratedClaimerLightDid = LightDidDetails.fromIdentifier(
-      encodeAddress(unmigratedClaimerKey.keypair.publicKey, 38),
-      VerificationKeyType.Sr25519
-    )
-    const migratedClaimerKey = makeSigningKeyTool()
-    migratedClaimerLightDid = LightDidDetails.fromIdentifier(
-      encodeAddress(migratedClaimerKey.keypair.publicKey, 38),
-      VerificationKeyType.Sr25519
-    )
-    // Change also the authentication key of the full DID to properly verify signature verification,
-    // so that it uses a completely different key and the credential is still correctly verified.
-    newKeyForMigratedClaimerDid = makeSigningKeyTool()
-    migratedClaimerFullDid = await createMinimalFullDidFromLightDid(
-      migratedClaimerLightDid as LightDidDetails,
-      {
-        ...newKeyForMigratedClaimerDid.authenticationKey,
-        id: 'new-auth',
-      }
-    )
-    migratedThenDeletedKey = makeSigningKeyTool(SigningAlgorithms.Ed25519)
-    migratedThenDeletedClaimerLightDid = LightDidDetails.fromIdentifier(
-      encodeAddress(migratedThenDeletedKey.keypair.publicKey, 38),
-      VerificationKeyType.Ed25519
-    )
-    migratedThenDeletedClaimerFullDid = createMinimalFullDidFromLightDid(
-      migratedThenDeletedClaimerLightDid as LightDidDetails
-    )
-
-    const rawCType: ICType['schema'] = {
-      $id: 'kilt:ctype:0x1',
-      $schema: 'http://kilt-protocol.org/draft-01/ctype#',
-      title: 'credential',
-      properties: {
-        name: { type: 'string' },
-      },
-      type: 'object',
-    }
-
-    ctype = CType.fromSchema(rawCType, migratedClaimerFullDid.uri)
-
-    // cannot be used since the variable needs to be established in the outer scope
-    reqForAtt = Credential.fromClaim(
-      Claim.fromCTypeAndClaimContents(
-        ctype,
-        {
-          name: 'Peter',
-          age: 12,
-        },
-        migratedClaimerFullDid.uri
-      )
-    )
-  })
-
-  it('should create presentation and exclude specific attributes using a full DID', async () => {
-    const challenge = UUID.generate()
-    const presentation = await Credential.createPresentation({
-      credential: reqForAtt,
-      selectedAttributes: ['name'],
-      sign: newKeyForMigratedClaimerDid.sign,
-      claimerDid: migratedClaimerFullDid,
-      challenge,
-    })
-    await expect(
-      Credential.verify(presentation, {
-        resolver: mockResolver,
-      })
-    ).resolves.not.toThrow()
-    expect(presentation.claimerSignature?.challenge).toEqual(challenge)
-  })
-  it('should create presentation and exclude specific attributes using a light DID', async () => {
-    const rawCType: ICType['schema'] = {
-      $id: 'kilt:ctype:0x1',
-      $schema: 'http://kilt-protocol.org/draft-01/ctype#',
-      title: 'credential',
-      properties: {
-        name: { type: 'string' },
-      },
-      type: 'object',
-    }
-    ctype = CType.fromSchema(rawCType, attester.uri)
-
-    // cannot be used since the variable needs to be established in the outer scope
-    reqForAtt = Credential.fromClaim(
-      Claim.fromCTypeAndClaimContents(
-        ctype,
-        {
-          name: 'Peter',
-          age: 12,
-        },
-        unmigratedClaimerLightDid.uri
-      )
-    )
-
-    const challenge = UUID.generate()
-    const presentation = await Credential.createPresentation({
-      credential: reqForAtt,
-      selectedAttributes: ['name'],
-      sign: unmigratedClaimerKey.sign,
-      claimerDid: unmigratedClaimerLightDid,
-      challenge,
-    })
-    await expect(
-      Credential.verify(presentation, {
-        resolver: mockResolver,
-      })
-    ).resolves.not.toThrow()
-    expect(presentation.claimerSignature?.challenge).toEqual(challenge)
-  })
-  it('should create presentation and exclude specific attributes using a migrated DID', async () => {
-    const rawCType: ICType['schema'] = {
-      $id: 'kilt:ctype:0x1',
-      $schema: 'http://kilt-protocol.org/draft-01/ctype#',
-      title: 'credential',
-      properties: {
-        name: { type: 'string' },
-      },
-      type: 'object',
-    }
-    ctype = CType.fromSchema(rawCType, attester.uri)
-
-    // cannot be used since the variable needs to be established in the outer scope
-    reqForAtt = Credential.fromClaim(
-      Claim.fromCTypeAndClaimContents(
-        ctype,
-        {
-          name: 'Peter',
-          age: 12,
-        },
-        // Use of light DID in the claim.
-        migratedClaimerLightDid.uri
-      )
-    )
-
-    const challenge = UUID.generate()
-    const presentation = await Credential.createPresentation({
-      credential: reqForAtt,
-      selectedAttributes: ['name'],
-      sign: newKeyForMigratedClaimerDid.sign,
-      // Use of full DID to sign the presentation.
-      claimerDid: migratedClaimerFullDid,
-      challenge,
-    })
-    await expect(
-      Credential.verify(presentation, {
-        resolver: mockResolver,
-      })
-    ).resolves.not.toThrow()
-    expect(presentation.claimerSignature?.challenge).toEqual(challenge)
-  })
-
-  it('should fail to create a valid presentation and exclude specific attributes using a light DID after it has been migrated', async () => {
-    const rawCType: ICType['schema'] = {
-      $id: 'kilt:ctype:0x1',
-      $schema: 'http://kilt-protocol.org/draft-01/ctype#',
-      title: 'credential',
-      properties: {
-        name: { type: 'string' },
-      },
-      type: 'object',
-    }
-    ctype = CType.fromSchema(rawCType, attester.uri)
-
-    // cannot be used since the variable needs to be established in the outer scope
-    reqForAtt = Credential.fromClaim(
-      Claim.fromCTypeAndClaimContents(
-        ctype,
-        {
-          name: 'Peter',
-          age: 12,
-        },
-        // Use of light DID in the claim.
-        migratedClaimerLightDid.uri
-      )
-    )
-
-    const challenge = UUID.generate()
-    const att = await Credential.createPresentation({
-      credential: reqForAtt,
-      selectedAttributes: ['name'],
-      sign: newKeyForMigratedClaimerDid.sign,
-      // Still using the light DID, which should fail since it has been migrated
-      claimerDid: migratedClaimerLightDid,
-      challenge,
-    })
-    await expect(
-      Credential.verify(att, {
-        resolver: mockResolver,
-      })
-    ).resolves.toBeFalsy()
-  })
-
-  it('should fail to create a valid presentation using a light DID after it has been migrated and deleted', async () => {
-    const rawCType: ICType['schema'] = {
-      $id: 'kilt:ctype:0x1',
-      $schema: 'http://kilt-protocol.org/draft-01/ctype#',
-      title: 'credential',
-      properties: {
-        name: { type: 'string' },
-      },
-      type: 'object',
-    }
-    ctype = CType.fromSchema(rawCType, attester.uri)
-
-    // cannot be used since the variable needs to be established in the outer scope
-    reqForAtt = Credential.fromClaim(
-      Claim.fromCTypeAndClaimContents(
-        ctype,
-        {
-          name: 'Peter',
-          age: 12,
-        },
-        // Use of light DID in the claim.
-        migratedThenDeletedClaimerLightDid.uri
-      )
-    )
-
-    const challenge = UUID.generate()
-    const presentation = await Credential.createPresentation({
-      credential: reqForAtt,
-      selectedAttributes: ['name'],
-      sign: migratedThenDeletedKey.sign,
-      // Still using the light DID, which should fail since it has been migrated and then deleted
-      claimerDid: migratedThenDeletedClaimerLightDid,
-      challenge,
-    })
-    await expect(
-      Credential.verify(presentation, {
-        resolver: mockResolver,
-      })
-    ).resolves.toBeFalsy()
-  })
-
-  it('should verify the credential claims structure against the ctype', () => {
-    expect(Credential.verifyAgainstCType(reqForAtt, ctype)).toBeTruthy()
-    reqForAtt.claim.contents.name = 123
-
-    expect(Credential.verifyAgainstCType(reqForAtt, ctype)).toBeFalsy()
   })
 })
